Update HubTemperature.ino
There may be a more elegant way to do it, but I needed to hide the SSID (AI-Thinker) that was being broadcast from the device. The combination of adding "AT+CWSAP?" and changing "AT+CWMODE=1" seemed to do the trip. Device is connected and broadcasting without showing up as separate AP.